vault-service/requirements.txt
boto3>=1.34.0
pypdf>=3.17.4
python-docx>=1.1.0
openai>=1.6.1
python-multipart>=0.0.6
pinecone-client>=5.0.0
